Split masking policies into files
