+++ pkg/sdk/masking_policies_ext.go
@@ -1,0 +1,17 @@
+package sdk
+
+import "encoding/json"
+
+type MaskingPolicyOptions struct {
+	ExemptOtherPolicies bool `json:"EXEMPT_OTHER_POLICIES"`
+}
+
+func ParseMaskingPolicyOptions(s string) (MaskingPolicyOptions, error) {
+	var options MaskingPolicyOptions
+	err := json.Unmarshal([]byte(s), &options)
+	if err != nil {
+		return MaskingPolicyOptions{}, err
+	}
+
+	return options, nil
+}
